requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
bson==0.5.10
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.2
curl_cffi==0.13.0
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.6.1
et_xmlfile==2.0.0
exchangelib==5.4.2
Faker==37.1.0
fastjsonschema==2.20.0
fonttools==4.44.3
frozendict==2.4.6
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
idna==3.6
isodate==0.6.1
kiwisolver==1.4.5
lxml==5.2.2
markdown==3.8.2
matplotlib==3.8.1
msal==1.30.0
multitasking==0.0.12
networkx==3.1
numpy>=1.26.0
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==23.2
pandas>=2.2.3
passlib==1.7.4
peewee==3.18.2
Pillow==10.0.1
platformdirs==4.3.8
ply==3.11
protobuf>=6.0
pure_eval==0.2.3
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PySocks==1.7.1
pyspnego==0.11.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.31.0
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sspilib==0.1.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.13
wsproto==1.2.0
yfinance==0.2.65
z3-solver==4.13.0.0
